requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# Parsing & text
pypdf>=5.0.0
beautifulsoup4>=4.12.0

# IR
rank-bm25>=0.2.2
faiss-cpu>=1.8.0

# Embeddings & NLP
sentence-transformers>=3.0.0
spacy>=3.7.2

# Optional: neural QG (falls back to templates if unavailable)
transformers>=4.42.0
torch>=2.3.0

# App
streamlit>=1.36.0
